Fixed a size_t uint conversion
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@558 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Core/Mesh/PolyConnectivity.cc
@@ -111,7 +111,7 @@ void PolyConnectivity::adjust_outgoing_halfedge(VertexHandle _vh)
 //-----------------------------------------------------------------------------
 
 PolyConnectivity::FaceHandle
-PolyConnectivity::add_face(const VertexHandle* _vertex_handles, uint _vhs_size)
+PolyConnectivity::add_face(const VertexHandle* _vertex_handles, size_t _vhs_size)
 {
   VertexHandle                   vh;
   uint                           i, ii, n(_vhs_size), id;
